requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
Pillow==10.2.0
PyPDF2==3.0.1
spacy==3.7.2
openai==1.12.0

